+++ src/setuplib.py
@@ -1,9 +1,9 @@
-"""nv_plugin installer library module. 
+"""nv_zim installer library module. 
 
 Version @release
 
 Copyright (c) 2024 Peter Triesberger
-For further information see https://github.com/peter88213/nv_plugin
+For further information see https://github.com/peter88213/nv_zim
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 from shutil import copytree
@@ -18,7 +18,7 @@ except ModuleNotFoundError:
     print('The tkinter module is missing. Please install the tk support package for your python3 version.')
     sys.exit(1)
 
-PLUGIN = 'nv_plugin.py'
+PLUGIN = 'nv_zim.py'
 VERSION = ' @release'
 
 root = Tk()