+++ tools/build.py
@@ -1,9 +1,9 @@
-"""Build the nv_plugin novelibre plugin package.
+"""Build the nv_zim novelibre plugin package.
         
 Note: VERSION must be updated manually before starting this script.
 
 Copyright (c) 2024 Peter Triesberger
-For further information see https://github.com/peter88213/nv_plugin
+For further information see https://github.com/peter88213/nv_zim
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 import os
@@ -17,9 +17,9 @@ VERSION = '0.1.0'
 
 class PluginBuilder(PackageBuilder):
 
-    PRJ_NAME = 'nv_plugin'
-    LOCAL_LIB = 'nvpluginlib'
-    GERMAN_TRANSLATION = False
+    PRJ_NAME = 'nv_zim'
+    LOCAL_LIB = 'nvzimlib'
+    GERMAN_TRANSLATION = True
 
 
 def main():